+++ src/consent-manager/translations-utils.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/camelcase */
 import {
   LanguageData,
   TranslationsDictionary,